models/inventory-entry/src/inventory-entry-draft/presets/sample-data-b2c-lifestyle/sku-mmst-01.ts
import {
  ProductVariantDraft,
  type TProductVariantDraft,
} from '@commercetools-test-data/product';
import type { TInventoryEntryDraftBuilder } from '../../../types';
import * as InventoryEntryDraft from '../../index';

const variant = ProductVariantDraft.presets.sampleData
  .minimalistModernSideTable01()
  .build<TProductVariantDraft>();

const skuMmst01 = (): TInventoryEntryDraftBuilder =>
  InventoryEntryDraft.presets.empty().sku(variant.sku!).quantityOnStock(72);
export default skuMmst01;
